Refactor : Added Auth & Goto URL
