+++ app/constants/signup.js
@@ -1,8 +1,3 @@
-export const GOTO_URL = 'https://realdevsquad.com/goto';
-
-export const AUTH_URL =
-  'https://github.com/login/oauth/authorize?client_id=23c78f66ab7964e5ef97';
-
 export const REDIRECT_TEXT = {
   loggedIn: 'You have not logged in. Please login first to fill this form.',
   formAlreadyFilled:
