+++ app/controllers/signup.js
@@ -2,7 +2,7 @@ import Controller from '@ember/controller';
 import { action, set } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
-import { GOTO_URL } from '../constants/signup';
+import { GOTO_URL } from '../constants/url';
 import { OLD_SIGNUP_FLOW } from '../constants/analytics';
 import { toastNotificationTimeoutOptions } from '../constants/toast-notification';
 import ENV from 'website-my/config/environment'; // remove this when new flow goes live
